CM - Adição herança nos usuários, ajuste de pastas e nomes e injeção em testes unitários
