+++ sme.gclass.api.domain/Entidades/Google/Usuarios/Alunos/AlunoGoogle.cs
@@ -1,0 +1,17 @@
+﻿namespace SME.GoogleClassroom.Dominio
+{
+    public class AlunoGoogle : UsuarioGoogle
+    {
+        public int Codigo { get; set; }
+
+        public AlunoGoogle(int codigo, string nome, string email, string organizationPath)
+            : base(nome, email, organizationPath)
+        {
+            Codigo = codigo;
+        }
+
+        protected AlunoGoogle()
+        {
+        }
+    }
+}
