+++ sme.gclass.api.domain/Entidades/Google/Usuarios/Alunos/Cursos/AlunoCursoGoogle.cs
