+++ sme.gclass.api.domain/Entidades/Google/Usuarios/Cursos/CursoUsuario.cs
