+++ sme.gclass.api.domain/Entidades/Google/Usuarios/Funcionarios/FuncionarioGoogle.cs
@@ -1,0 +1,17 @@
+﻿namespace SME.GoogleClassroom.Dominio
+{
+    public class FuncionarioGoogle : UsuarioGoogle
+    {
+        public long Rf { get; set; }
+
+        public FuncionarioGoogle(long rf, string nome, string email, string organizationPath)
+            : base(nome, email, organizationPath)
+        {
+            Rf = rf;
+        }
+
+        protected FuncionarioGoogle()
+        {
+        }
+    }
+}
