+++ sme.gclass.api.domain/Entidades/Google/Usuarios/Professores/Cursos/ProfessorCursoGoogle.cs
