+++ sme.gclass.api.domain/Entidades/Google/Usuarios/UsuarioGoogle.cs
@@ -2,13 +2,64 @@
 
 namespace SME.GoogleClassroom.Dominio
 {
-    public class UsuarioGoogle
+    public abstract class UsuarioGoogle
     {
-        public string Id { get; set; }
-        public string Nome { get; set; }
+        private const int TamanhoMaximoDoSobrenome = 60;
+
+        public long Indice { get; set; }
+
+        private string nome;
+
+        public string Nome
+        {
+            get => nome;
+            set => SetNome(value);
+        }
+
+        public string PrimeiroNome
+        {
+            get
+            {
+                if (string.IsNullOrWhiteSpace(nome)) return string.Empty;
+                return nome.Split(' ')[0];
+            }
+        }
+
+        public string Sobrenome
+        {
+            get
+            {
+                if (string.IsNullOrWhiteSpace(nome)) return string.Empty;
+                return nome.Replace(PrimeiroNome, string.Empty).Trim();
+            }
+        }
+
         public string Email { get; set; }
         public string OrganizationPath { get; set; }
-        public bool? Suspenso { get; set; }       
-        public DateTime? DataCriacao { get; set; }
+        public DateTime DataInclusao { get; set; }
+        public DateTime? DataAtualizacao { get; set; }
+
+        public UsuarioGoogle(string nome, string email, string organizationPath)
+        {
+            Nome = nome;
+            Email = email;
+            OrganizationPath = organizationPath;
+            DataInclusao = DateTime.Now;
+        }
+
+        protected UsuarioGoogle()
+        {
+        }
+
+        private void SetNome(string nome)
+        {
+            this.nome = nome;
+            if (Sobrenome.Length <= TamanhoMaximoDoSobrenome)
+                return;
+
+            var sobrenomeTrucado = Sobrenome.Substring(0, TamanhoMaximoDoSobrenome);
+            var nomeFormatado = nome.Replace(Sobrenome, sobrenomeTrucado).Trim(); ;
+            this.nome = nomeFormatado;
+        }
     }
 }
