+++ sme.gclass.application/Queries/UsuarioGoogle/ObterUsuarioGoogle/ObterUsuarioGoogleQueryHandler.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 
 namespace SME.GoogleClassroom.Aplicacao
 {
-    public class ObterUsuarioGoogleQueryHandler : IRequestHandler<ObterUsuarioGoogleQuery, UsuarioGoogle>
+    public class ObterUsuarioGoogleQueryHandler : IRequestHandler<ObterUsuarioGoogleQuery, UsuarioGoogleDto>
     {
         private readonly IMediator mediator;
         private readonly IAsyncPolicy policy;
@@ -21,20 +21,20 @@ namespace SME.GoogleClassroom.Aplicacao
             this.policy = registry.Get<AsyncRetryPolicy>("RetryPolicy");
         }
 
-        public async Task<UsuarioGoogle> Handle(ObterUsuarioGoogleQuery request, CancellationToken cancellationToken)
+        public async Task<UsuarioGoogleDto> Handle(ObterUsuarioGoogleQuery request, CancellationToken cancellationToken)
         {
             var diretorioClassroom = await mediator.Send(new ObterDirectoryServiceGoogleClassroomQuery());
             return await policy.ExecuteAsync(() => ObterUsuarioNoGoogle(request.Email, diretorioClassroom));
 
         }
 
-        private async Task<UsuarioGoogle> ObterUsuarioNoGoogle(string email, DirectoryService diretorioClassroom)
+        private async Task<UsuarioGoogleDto> ObterUsuarioNoGoogle(string email, DirectoryService diretorioClassroom)
         {
 
             var requestCreate = diretorioClassroom.Users.Get(email);
             var usuario = await requestCreate.ExecuteAsync();            
 
-            return new UsuarioGoogle()
+            return new UsuarioGoogleDto()
             {
                 Id = usuario.Id,
                 Nome = usuario.Name.FullName,
