+++ sme.gclass.application/SME.GoogleClassroom.Aplicacao.csproj
@@ -5,6 +5,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Queries\Execucao\**" />
+    <EmbeddedResource Remove="Queries\Execucao\**" />
+    <None Remove="Queries\Execucao\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="FluentValidation" Version="9.5.1" />
     <PackageReference Include="Google.Apis.Admin.Directory.directory_v1" Version="1.49.0.1982" />
     <PackageReference Include="Google.Apis.Classroom.v1" Version="1.50.0.2223" />
@@ -18,9 +24,5 @@
     <ProjectReference Include="..\sme.gclass.api.dados\SME.GoogleClassroom.Dados.csproj" />
     <ProjectReference Include="..\sme.gclass.api.domain\SME.GoogleClassroom.Dominio.csproj" />
     <ProjectReference Include="..\sme.gclass.api.infra\SME.GoogleClassroom.Infra.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Queries\Execucao\" />
   </ItemGroup>
 </Project>
